requirements.txt
fastapi
uvicorn
wikipedia-api
